+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 + [x] Added config parameter `precision`, now you control balance precision, default `4`
 + [x] Config parameters `price:*`, `intervals:*`, `blockDiff`, `timesToSwitchNode` deprecated
 + [x] To set using host via `HTTPS/WSS` use parameter `https` instead of `secure` in `nodes`
++ [x] Added supercharge info to rewards
 
 ### 1.0.4
 + [x] Increased chart size to 100 points